refactor: 디바이스 관련 api 호출 구조 변경
