+++ src/background.js
@@ -5,7 +5,7 @@
  * API 요청은 CORS 우회를 위해 이 서비스 워커에서 처리한다.
  */
 
-import { CONFIG } from './config.js';
+import { handleApiRequest } from './api-proxy.js';
 
 // ============================================
 // 익스텐션 설치/업데이트 이벤트
@@ -18,53 +18,7 @@ chrome.runtime.onInstalled.addListener((details) => {
   }
 });
 
-// ============================================
-// API 프록시 핸들러
-// ============================================
-async function handleApiRequest(request) {
-  const { endpoint, method = 'GET', body, params } = request;
 
-  let url = `${CONFIG.BASE_URL}${endpoint}`;
-  if (params) {
-    url += `?${new URLSearchParams(params)}`;
-  }
-
-  const options = {
-    method,
-    headers: { 'Content-Type': 'application/json' }
-  };
-
-  if (body) {
-    options.body = JSON.stringify(body);
-  }
-
-  try {
-    const response = await fetch(url, options);
-
-    // 204 No Content
-    if (response.status === 204) {
-      return { success: true, data: null };
-    }
-
-    let data;
-    try {
-      data = await response.json();
-    } catch (parseError) {
-      console.error('Failed to parse JSON response:', parseError);
-      data = { message: 'Invalid JSON response from server.' };
-    }
-
-    if (!response.ok) {
-      console.error('API request failed:', { url, status: response.status, data });
-      return { success: false, status: response.status, message: data.message };
-    }
-
-    return { success: true, data };
-  } catch (error) {
-    console.error('Network request failed:', { url, error });
-    return { success: false, status: 0, message: error.message, isNetworkError: true };
-  }
-}
 
 // ============================================
 // 메시지 리스너 (popup과 통신)
